# server/documentation/openapi.yaml
...
openapi: 3.1.0
info:
  title: Cosmian Authentication Verifier
  version: 1.0.0
  description: |
    High-performance authentication and session management server.
    All endpoints are served over **HTTPS only**.

    ## Authentication for admin endpoints
    Admin endpoints require a session cookie obtained by logging into the `_` realm:
    ```
    POST /login?realm=_
    Authorization: Basic <base64(username:password)>
    ```
    The server responds with `Set-Cookie: _ea_=<value>; HttpOnly; Secure; SameSite=Strict`.
    Pass this cookie on every subsequent admin request.
servers:
  - url: https://localhost:8443
    description: Local development
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: _ea_
    basicAuth:
      type: http
      scheme: basic
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    mtls:
      type: mutualTLS
    appToken:
      type: apiKey
      in: header
      name: X-Vault-Token
      description: |
        AppRole-compatible bearer token issued by this server's AppRole or
        Kubernetes login endpoints.  Required on `/auth/token/*` endpoints.
...
paths:
...
  /oidc/authorize/login:
    post:
      tags: [OIDC]
      summary: Submit login credentials (front-channel)
      operationId: oidcAuthorizeLogin
      description: |
        Authenticates the end-user against the client's realm userpass store
        (with TOTP when enabled) and renders the consent screen. CSRF-protected
        by the server-signed `flow_token`.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [flow_token, username, password]
              properties:
                flow_token: {type: string}
                username: {type: string}
                password: {type: string}
                totp_code: {type: string}
      responses:
        '200':
          description: HTML consent screen (or login form re-rendered on error).
          content:
            text/html:
              schema: {type: string}
  /oidc/authorize/consent:
    post:
      tags: [OIDC]
      summary: Submit the consent decision (front-channel)
      operationId: oidcAuthorizeConsent
      description: |
        On approval, issues a single-use authorization code bound to the PKCE
        challenge and redirects back to the client's `redirect_uri` with
        `code`/`state`. On denial, redirects with `error=access_denied`.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [flow_token, decision]
              properties:
                flow_token: {type: string}
                decision: {type: string, enum: [approve, deny]}
      responses:
        '302':
          description: Redirect back to the client with the code or an error.
...